# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workout Tracker API",
    "version": "1.0.0",
    "description": "API documentation for the Workout Tracker application"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/workouts": {
      "get": {
        "summary": "Get all workout plans",
        "responses": {
          "200": {
            "description": "Successful retrieval of workout plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutPlan"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No workout plan found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create a new workout plan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutPlan"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWorkoutPlan"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/workouts/{plan_id}": {
      "get": {
        "summary": "Get workout plan by ID",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "description": "ID of the workout plan to retrieve",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of workout plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWorkoutPlan"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update a workout plan",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "description": "ID of the workout plan to update",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutPlan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workout plan updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWorkoutPlan"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a workout plan",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "description": "ID of the workout plan to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workout plan deleted successfully"
          },
          "404": {
            "description": "Workout plan not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/workouts/{plan_id}/weeks": {
      "post": {
        "summary": "Add a week to a workout plan",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "description": "ID of the workout plan to add week to",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeek"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Week added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWeek"
                }
              }
            }
          },
          "400": {
            "description": "Invalid week number or week already exists"
          },
          "404": {
            "description": "Workout plan not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/workouts/weeks/{week_id}/days": {
      "post": {
        "summary": "Add a day to a workout week",
        "parameters": [
          {
            "name": "week_id",
            "in": "path",
            "description": "ID of the workout week to add day to",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutDay"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Day added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWorkoutDay"
                }
              }
            }
          },
          "400": {
            "description": "Day already exists in this week"
          },
          "404": {
            "description": "Week not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/workouts/days/{day_id}/exercises": {
      "get": {
        "summary": "Get exercises in a workout day",
        "parameters": [
          {
            "name": "day_id",
            "in": "path",
            "description": "ID of the workout day to retrieve exercises for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of exercises",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutPlanExercise"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Day not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Add an exercise to a workout day",
        "parameters": [
          {
            "name": "day_id",
            "in": "path",
            "description": "ID of the workout day to add exercise to",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddExerciseToWorkout"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercise added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWorkoutPlanExercise"
                }
              }
            }
          },
          "404": {
            "description": "Day or Exercise not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update exercises in a workout day",
        "parameters": [
          {
            "name": "day_id",
            "in": "path",
            "description": "ID of the workout day to update exercises for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UpdateExerciseInWorkout"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercises updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutPlanExercise"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Day or Exercise not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/workouts/days/{day_id}/exercises/{exercise_id}": {
      "delete": {
        "summary": "Delete an exercise from a workout day",
        "parameters": [
          {
            "name": "day_id",
            "in": "path",
            "description": "ID of the workout day",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "exercise_id",
            "in": "path",
            "description": "ID of the exercise to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Exercise deleted successfully"
          },
          "404": {
            "description": "Exercise not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateWorkoutPlan": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "weeks": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": ["name"]
      },
      "DisplayWorkoutPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "user_id", "name", "created_at", "updated_at"]
      },
      "UpdateWorkoutPlan": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateWeek": {
        "type": "object",
        "properties": {
          "week_number": {
            "type": "integer"
          }
        },
        "required": ["week_number"]
      },
      "DisplayWeek": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "week_number": {
            "type": "integer"
          }
        },
        "required": ["id", "week_number"]
      },
      "CreateWorkoutDay": {
        "type": "object",
        "properties": {
          "day_of_week": {
            "type": "string"
          }
        },
        "required": ["day_of_week"]
      },
      "DisplayWorkoutDay": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "day_of_week": {
            "type": "string"
          }
        },
        "required": ["id", "day_of_week"]
      },
      "AddExerciseToWorkout": {
        "type": "object",
        "properties": {
          "exercise_id": {
            "type": "integer"
          },
          "workout_plan_day_id": {
            "type": "integer"
          },
          "sets": {
            "type": "integer"
          },
          "reps": {
            "type": "integer"
          },
          "order": {
            "type": "integer"
          }
        },
        "required": ["exercise_id", "workout_plan_day_id", "sets", "reps", "order"]
      },
      "UpdateExerciseInWorkout": {
        "type": "object",
        "properties": {
          "exercise_id": {
            "type": "integer"
          },
          "sets": {
            "type": "integer",
            "nullable": true
          },
          "reps": {
            "type": "integer",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": ["exercise_id"]
      },
      "DisplayWorkoutPlanExercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "workout_plan_day_id": {
            "type": "integer"
          },
          "exercise_id": {
            "type": "integer"
          },
          "sets": {
            "type": "integer"
          },
          "reps": {
            "type": "integer"
          },
          "order": {
            "type": "integer"
          }
        },
        "required": ["id", "workout_plan_day_id", "exercise_id", "sets", "reps", "order"]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}